requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2
django-rest-captcha==0.3.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
model-mommy==2.0.0
peppercorn==0.6
Pillow==10.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
